Fixed a number of warnings for VS.

+++ runtime/Cpp/runtime/src/ParserInterpreter.h
@@ -157,7 +157,7 @@ namespace antlr4 {
      *  Those values are used to create new recursive rule invocation contexts
      *  associated with left operand of an alt like "expr '*' expr".
      */
-    std::stack<std::pair<Ref<ParserRuleContext>, int>> _parentContextStack;
+    std::stack<std::pair<Ref<ParserRuleContext>, size_t>> _parentContextStack;
     
     /** We need a map from (decision,inputIndex)->forced alt for computing ambiguous
      *  parse trees. For now, we allow exactly one override.
