+++ runtime/Cpp/runtime/src/atn/LookaheadEventInfo.cpp
@@ -34,8 +34,8 @@
 using namespace antlr4;
 using namespace antlr4::atn;
 
-LookaheadEventInfo::LookaheadEventInfo(int decision, ATNConfigSet *configs, int predictedAlt,
-  TokenStream *input, int startIndex, int stopIndex, bool fullCtx)
+LookaheadEventInfo::LookaheadEventInfo(size_t decision, ATNConfigSet *configs, size_t predictedAlt,
+  TokenStream *input, size_t startIndex, size_t stopIndex, bool fullCtx)
   : DecisionEventInfo(decision, configs, input, startIndex, stopIndex, fullCtx) {
     
   this->predictedAlt = predictedAlt;
