+++ runtime/Cpp/runtime/src/atn/LookaheadEventInfo.h
@@ -44,7 +44,7 @@ namespace atn {
     ///  the outermost alt shown for a rule; left-recursive rules have
     ///  user-level alts that differ from the rewritten rule with a (...) block
     ///  and a (..)* loop.
-    int predictedAlt = 0;
+    size_t predictedAlt = 0;
 
     /// <summary>
     /// Constructs a new instance of the <seealso cref="LookaheadEventInfo"/> class with
@@ -60,8 +60,8 @@ namespace atn {
     /// <param name="fullCtx"> {@code true} if the current lookahead is part of an LL
     /// prediction; otherwise, {@code false} if the current lookahead is part of
     /// an SLL prediction </param>
-    LookaheadEventInfo(int decision, ATNConfigSet *configs, int predictedAlt, TokenStream *input, int startIndex,
-                       int stopIndex, bool fullCtx);
+    LookaheadEventInfo(size_t decision, ATNConfigSet *configs, size_t predictedAlt, TokenStream *input, size_t startIndex,
+                       size_t stopIndex, bool fullCtx);
   };
 
 } // namespace atn
